% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{annotation.map}
\alias{annotation.map}
\title{Map annotation to gene sets and genes}
\usage{
annotation.map(annotation, gs_annos, gs_genes)
}
\arguments{
\item{annotation}{annotation to map}

\item{gs_annos}{value from \code{\link{process.annotations}}}

\item{gs_genes}{value from \code{\link{process.database}}}
}
\value{
vector of associated gene sets and vector of associated genes
}
\description{
Map annotation to gene sets and genes
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
anno.pre <- import.annotations(anno.raw, c(2, 9), 10)
info.pre <- anno.pre[c('name', 'info')]
anno <- process.annotations(anno.pre, info.pre, c('file', 'auto'))

data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)
data <- process.database(data.pre, 'Not assigned', 'Not assigned')

annotation.map('annotation', anno$gs_annos, data$gs_genes)
}
}
