% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{annotation_map}
\alias{annotation_map}
\title{Map annotation to gene sets and genes}
\usage{
annotation_map(annotation, gs_annos, gs_genes)
}
\arguments{
\item{annotation}{annotation to map}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}
}
\value{
vector of associated gene sets and vector of associated genes
}
\description{
Map annotation to gene sets and genes
}
\examples{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
data <- process_database(data_pre, 'Not assigned', 'Not assigned')
annotation_map("Carcinogen", anno$gs_annos, data$gs_genes)
}
