% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate.post}
\alias{calculate.post}
\title{Finish statistical calculations}
\usage{
calculate.post(stats.pre, input.count, universe)
}
\arguments{
\item{stats.pre}{value from \code{\link{calculate.pre}}}

\item{input.count}{number of genes in input}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Finish statistical calculations
}
\examples{
\dontrun{
calculate.popst(stats.pre, 100, 10000)
}
}
