% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate.pre}
\alias{calculate.pre}
\title{Pre-calculate overlap statistics}
\usage{
calculate.pre(input, annos, gs_annos, gs_genes)
}
\arguments{
\item{input}{output of \code{\link{process.input}}}

\item{annos}{value from \code{\link{process.annotations}}}

\item{gs_annos}{value from \code{\link{process.annotations}}}

\item{gs_genes}{value from \code{\link{process.database}}}
}
\value{
partially-computed statistics and list of annotation matches
}
\description{
Pre-calculate overlap statistics
}
\examples{
\dontrun{
calculate.pre(input, annos, gs_annos, gs_genes)
}
}
