% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, annos, gs_annos, gs_genes, universe)
}
\arguments{
\item{input}{output of \code{\link{process.input}}}

\item{annos}{value from \code{\link{process.annotations}}}

\item{gs_annos}{value from \code{\link{process.annotations}}}

\item{gs_genes}{value from \code{\link{process.database}}}

\item{universe}{number of genes in universe}
}
\value{
tibble of overlap statistics and list of annotation matches
}
\description{
Calculate overlap statistics
}
\examples{
\dontrun{
calculate(input, annos, gs_annos, gs_genes, 10000)
}
}
