% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_post}
\alias{calculate_post}
\title{Finish statistical calculations}
\usage{
calculate_post(stats_pre, input_count, universe)
}
\arguments{
\item{stats_pre}{value from \code{\link{calculate_pre}}}

\item{input_count}{number of genes in input}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Finish statistical calculations
}
\examples{
\dontrun{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
calc_pre <- calculate_pre(input, anno$annos, anno$gs_annos, data$gs_genes)
calc <- calculate_post(calc_pre$stats_pre, nrow(input), 100)
}
}
