% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_pre}
\alias{calculate_pre}
\title{Pre-calculate overlap statistics}
\usage{
calculate_pre(input, annos, gs_annos, gs_genes)
}
\arguments{
\item{input}{output of \code{\link{process_input}}}

\item{annos}{value from \code{\link{process_annotations}}}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}
}
\value{
partially-computed statistics and list of annotation matches
}
\description{
Pre-calculate overlap statistics
}
\examples{
\dontrun{
anno_raw <- read_common('path/to/anno.csv', ',', T)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)
info_pre <- anno_pre[c('name', 'info')]
anno <- process_annotations(anno_pre, info_pre, c('file', 'auto'))

data_raw <- read_common('path/to/data.csv', ',', T)
data_pre <- import_database(data_raw, c(2, 9), 10)
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
calculate_pre(input, anno$annos, anno$gs_annos, data$gs_genes)
}
}
