% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compute}
\alias{compute}
\title{Perform computations on imported and processed data}
\usage{
compute(
  input,
  anno,
  data,
  universe = NULL,
  info_from = "database",
  anno_opts = "file",
  categories = FALSE,
  organisms = FALSE,
  save = NULL
)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{anno}{output of \code{\link{import_annotations}}}

\item{data}{output of \code{\link{import_database}} or
\code{\link{import_msigdb}}}

\item{universe}{number of genes in universe}

\item{info_from}{optional: \code{"annotation"} or (default) \code{"database"}}

\item{anno_opts}{optional: \code{"name"} for gene set names, \code{"syms"}
for gene set symbols, \code{"info"} for gene set descriptions,
\code{"auto"} for automatically generated annotations and/or (default)
\code{"file"} for manual annotations}

\item{categories}{optional: categories to include; default all}

\item{organisms}{optional: organisms to include; default all}

\item{save}{optional: path to save overlap statistics as \code{.csv}}
}
\value{
\code{stats} tibble: overlap statistics

\code{matches} list: names: annotations vector: matched genes
}
\description{
Perform computations on imported and processed data
}
\examples{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", FALSE, c(2, 10), 11)
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

input <- process_input_text('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
results <- compute(input, anno, data, 10000)
}
