% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display.overlap}
\alias{display.overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
display.overlap(input, stats, matches, value, value.trans = "identity")
}
\arguments{
\item{input}{value from \code{\link{process.input}}}

\item{stats}{value from \code{\link{calculate}}}

\item{matches}{value from \code{\link{calculate}}}

\item{value}{option for tile color as 'Gene Value' or one of '# gene sets',
'#genes', '# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio',
'Fold Enrichment' or 'Adjusted Fold Enrichment'}

\item{value.trans}{optional value transformation}
}
\value{
ggplot heatmap
}
\description{
Plot gene-annotation matches as heatmap
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
anno.pre <- import.annotations(anno.raw, c(2, 9), 10)
info.pre <- anno.pre[c('name', 'info')]
anno <- process.annotations(anno.pre, info.pre, c('file', 'auto'))

data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)
data <- process.database(data.pre, 'Not assigned', 'Not assigned')

input <- process.input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
stats <- calculate(input, anno$annos, anno$gs_annos, data$gs_genes, 10000)
display.overlap(input$gene, stats$stats, stats$matches, 'Gene Value')
}
}
