% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_overlap}
\alias{display_overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
display_overlap(input, stats, matches, value, value_trans = "identity")
}
\arguments{
\item{input}{value from \code{\link{process_input}}}

\item{stats}{value from \code{\link{calculate}}}

\item{matches}{value from \code{\link{calculate}}}

\item{value}{option for tile color as 'Gene Value' or one of '# gene sets',
'#genes', '# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio',
'Fold Enrichment' or 'Adjusted Fold Enrichment'}

\item{value_trans}{optional value transformation}
}
\value{
ggplot heatmap
}
\description{
Plot gene-annotation matches as heatmap
}
\examples{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')$input
calc <- calculate(input, anno$annos, anno$gs_annos, data$gs_genes, 100)
over <- display_overlap(input, calc$stats, calc$matches, "Gene Value")
}
