% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_stats}
\alias{display_stats}
\title{Plot overlap statistics on a bar graph}
\usage{
display_stats(
  stats,
  value,
  color,
  value_trans = "identity",
  color_trans = "identity",
  sort = F
)
}
\arguments{
\item{stats}{value from \code{\link{calculate}}}

\item{value}{option for bar length as one of '# gene sets', '#genes',
'# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio', 'Fold Enrichment'
or 'Adjusted Fold Enrichment'}

\item{color}{option for bar color as one of '# gene sets', '#genes',
'# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio', 'Fold Enrichment'
or 'Adjusted Fold Enrichment'}

\item{value_trans}{optional value transformation}

\item{color_trans}{optional color transformation}

\item{sort}{whether to sort annotations alphabetically}
}
\value{
ggplot bar graph
}
\description{
Plot overlap statistics on a bar graph
}
\examples{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')$input
calc <- calculate(input, anno$annos, anno$gs_annos, data$gs_genes, 100)
stat <- display_stats(calc$stats, 'Fold Enrichment', 'Adjusted P-value')
}
