% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{explore_annotation}
\alias{explore_annotation}
\title{Find an annotation's associated gene sets and genes}
\usage{
explore_annotation(annotation, gs_annos, gs_genes, genes = NULL)
}
\arguments{
\item{annotation}{annotation to explore}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}

\item{genes}{optional: genes to match, or (default) all}
}
\value{
\code{"names"} vector: gene set names

\code{"genes"} vector: genes
}
\description{
Find an annotation's associated gene sets and genes
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", TRUE, c(2, 4), 5)
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)
info <- anno[c("name", "info")]

anno_proc <- process_annotations(anno, info, "file")
data_proc <- process_database(data, "Not assigned", "Not assigned")
anno_assoc <- explore_annotation("Carcinogen", anno_proc$gs_annos,
                                 data_proc$gs_genes)
}
}
\keyword{internal}
