% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacier.R
\docType{data}
\name{glacier}
\alias{glacier}
\title{Glaciers Around the World}
\format{A data frame with 132890 rows and 44 variables:
\describe{
  \item{wgi_glacier_id}{A 12-character unique glacier identifier. The ID number is assigned to the glacier as defined by the WGMS convention that forms the glacier ID number by combining the five following elements: 2-character political unit, 1-digit continent code, 4-character drainage code, 2-digit free position code, and 3-digit local glacier code.}
  \item{country_name}{A character string for which country or territory glacier is on.}
  \item{continent_name}{A character string for which continent glacier is on.}
  \item{glacier_name}{A character string for the name of the glacier. Up to 30 letters.}
  \item{lat}{The latitude of the glacier in decimal degrees North or South; up to 7 digits. Positive values indicate the Northern Hemisphere and negative values indicate the Southern Hemisphere. Latitude is given to a maximum precision of 4 decimal places.}
  \item{lon}{The longitude of the glacier in decimal degrees East or West; up to 7 digits. Positive values indicate east of the zero meridian and negative values indicate west of the zero meridian. Longitude is given to a maximum precision of 4 decimal places.}
  \item{easting}{Easting of local coordinate in UTM or other nationally determined format, up to 9 digits.}
  \item{northing}{Northing of local coordinates in UTM or other nationally determined format, up to 9 digits.}
  \item{num_basins}{The number of basins the glacier drains into.}
  \item{topo_year}{The 4-digit year of the topographic map used for measurements of glacier elevation. If multiple maps, used, most relevant included.}
  \item{topo_scale}{The scale of the topographic map used for measurements of glacier parameters, up to 7 digits. If multiple scales, used, most relevant included.}
  \item{photo_year}{The 4-digit year of the photograph used for measurements of glacier parameters. If multiple photographs, used, most relevant included.}
  \item{max_elev}{Maximum elevation of the highest point of the glacier, measured in meters above sea level, up to 4 digits.}
  \item{mean_elev}{Altitude of the contour line that halves the area of the glacier, measured in meters above sea level, up to 4 digits.}
  \item{min_elev}{The minimum elevation of the lowest point of the glacier in meters above sea level, up to 4 digits.}
  \item{min_elev_exp}{Minimum elevation exposed is the altitude of the lowest point of the total surface area of the glacier, in meters above sea level, that is not covered with coarse stone material, up to 4 digits.}
  \item{mean_elev_acc}{Mean elevation accumulation is the altitude of the contour line, in meters above sea level, that halves the accumulation area of the glacier, up to 4 digits.}
  \item{mean_elev_abl}{Mean elevation ablation is the altitude of the contour line, in meters above sea level, that halves the ablation area of the glacier, up to 4 digits.}
  \item{primary_class}{The primary classification of the glacier.}
  \item{form}{The form of the glacier.}
  \item{frontal_char}{The frontal characteristics of the glacier.}
  \item{longi_profile}{The longitudinal profile of the glacier.}
  \item{source_nourish}{The source of nourishment for the glacier.}
  \item{tongue_activity}{The activity of the tongue of the glacier.}
  \item{moraines1}{Moraines in contact with the present-day glacier.}
  \item{moraines2}{Moraines farther downstream of the glacier.}
  \item{period_activity_start}{4-digit start year of the period for which the tongue activity was assessed.}
  \item{period_activity_end}{4-digit end year of the period for which the tongue activity was assessed.}
  \item{snow_line_elev}{Altitude of the snow line of the glacier in meters above sea level, up to 4-digits.}
  \item{snow_line_acy}{Snow line accuracy rating in percentage range.}
  \item{snow_line_date}{8-digit date of observation of the snowline of the form YYYYMMDD where YYYY is the 4-digit year, MM is the 2-digit month, and DD is the 2-digit day of month. If part or all of the date is missing, that missing parts are filled with 9's. Note: Snow line elevation is the altitude of the transient snowline at the end of the ablation season or, in most cases and for practical reasons, at the time the photograph was taken.}
  \item{mean_depth}{The physical depth of the glacier in meters, up to 4 digits. Note I: Mean depth is only be given if actually measured (for instance by drilling or radio-echo soundings).}
  \item{depth_acy}{Depth accuracy rating in percentage range.}
  \item{total_area}{The total area of the glacier in a horizontal projection in square kilometers, up to 6 digits.}
  \item{area_acy}{Area accuracy rating of the total area in percentage range.}
  \item{area_in_state}{The total area of the glacier that resides in the political state concerned in a horizontal projection in square kilometers, up to 6 digits.}
  \item{area_exp}{The area of the exposed ice of the glacier in a horizontal projection in square kilometers, up to 6 digits}
  \item{mean_width}{The mean width of the glacier in a horizontal projection in kilometers, up to 4 digits.}
  \item{mean_length}{Mean length of the glacier in a horizontal projection in kilometers, up to 4 digits.}
  \item{max_length}{Maximum length of the glacier in kilometers measured along the most important flowline in a horizontal projection, up to 4 digits.}
  \item{max_length_exp}{Maximum length, in kilometers, of the exposed ice of the glacier in a horizontal projection, up to 4 digits.}
  \item{max_length_abl}{Maximum length, in kilometers, of the ablation area of the glacier in a horizontal projection, up to 4 digits.}
  \item{orientation_acc}{The 1- to 2-character main orientation of the accumulation area using the 8 cardinal points: N, NW, W, SW, S, SE, E, and NE.}
  \item{orientation_abl}{The 1- to 2-character main orientation of the ablation area using the 8 cardinal points: N, NW, W, SW, S, SE, E, and NE.}
 }
 @source \url{https://nsidc.org/data/G01130/versions/1}}
\usage{
glacier
}
\description{
A dataset of glaciers inventoried by the World Glacier Inventory (WGI). The data has been continuously collected since 1989 and was last updated in 2012 by the World Glacier Monitoring Service (WGMS).
}
\keyword{datasets}
