% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.annotations}
\alias{import.annotations}
\title{Import annotation file}
\usage{
import.annotations(anno.raw, cols = NULL, info = NULL)
}
\arguments{
\item{anno.raw}{raw file tibble or output of \code{\link{import.common}}}

\item{cols}{data column boundaries}

\item{info}{info column}
}
\value{
gene set annotations and information
}
\description{
Import annotation file
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
import.annotations(anno.raw, c(2, 9), 10)
}
}
