% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.database}
\alias{import.database}
\title{Import database file}
\usage{
import.database(common.raw, data = NULL, info = NULL)
}
\arguments{
\item{common.raw}{raw file tibble or output of \code{\link{import.common}}}

\item{data}{data column boundaries}

\item{info}{info column}
}
\value{
gene set contents and information
}
\description{
Import database file
}
\examples{
\dontrun{
import.database(raw, c(2, 9), 10)
}
}
