% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.msigdb_xml}
\alias{import.msigdb_xml}
\title{Import MSigDB XML file}
\usage{
import.msigdb_xml(path)
}
\arguments{
\item{path}{path to file}
}
\value{
gene set contents and information
}
\description{
Import MSigDB XML file
}
\examples{
\dontrun{
import.msigdb_xml('path/to/msigdb.xml')
}
}
