% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_annotations_file}
\alias{import_annotations_file}
\title{Process tibble of annotation file contents}
\usage{
import_annotations_file(anno_file, content, info)
}
\arguments{
\item{anno_file}{output of \code{\link{import_delim_path}}}

\item{content}{vector: first and last column of content}

\item{info}{column containing information}
}
\value{
glacier-specific imported annotations
}
\description{
Process tibble of annotation file contents
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_file <- import_delim_path(anno_path, ",", FALSE)
anno <- import_annotations_file(anno_file, c(2, 10), 11)
}
}
