% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_database}
\alias{import_database}
\title{Import delimited database file into glacier-specific format}
\usage{
import_database(path, delim, header, content = NULL, info = NULL)
}
\arguments{
\item{path}{path to file}

\item{delim}{file delimiter, usually \code{","} or \code{"\\t"}}

\item{header}{whether file contains header}

\item{content}{vector: first and last column of content; default 2 to end}

\item{info}{column containing information; default none}
}
\value{
glacier-specific imported database
}
\description{
Import delimited database file into glacier-specific format
}
\examples{
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)
}
