% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_database_file}
\alias{import_database_file}
\title{Process tibble of database file contents}
\usage{
import_database_file(data_file, content, info)
}
\arguments{
\item{data_file}{output of \code{\link{import_delim_path}}}

\item{content}{vector: first and last column of content}

\item{info}{column containing information}
}
\value{
glacier-specific imported database
}
\description{
Process tibble of database file contents
}
\examples{
\dontrun{
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
data_file <- import_delim_path(data_path, ",", FALSE)
data <- import_database_file(data_file, c(2, 4), 0)
}
}
