% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_delim_file}
\alias{import_delim_file}
\title{Process tibble of file contents}
\usage{
import_delim_file(file, content, info)
}
\arguments{
\item{file}{output of \code{\link{import_delim_path}}}

\item{content}{vector: first and last column of content}

\item{info}{column containing information}
}
\value{
tibble: renamed and cleaned file contents
}
\description{
Process tibble of file contents
}
\examples{
\dontrun{
path <- system.file("extdata", "ex_anno.csv", package = "glacier")
file <- import_delim_path(path, ",", FALSE)
data <- import_delim_file(file, c(2, 10), 11)
}
}
\keyword{internal}
