% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_delim_path}
\alias{import_delim_path}
\title{Read delimited files into tibble}
\usage{
import_delim_path(path, delim, header)
}
\arguments{
\item{path}{path to file}

\item{delim}{file delimiter, usually \code{","} or \code{"\\t"}}

\item{header}{whether file contains header}
}
\value{
tibble: file contents
}
\description{
Read delimited files into tibble
}
\examples{
\dontrun{
path <- system.file("extdata", "ex_anno.csv", package = "glacier")
file <- import_delim_path(path, ",", TRUE)
}
}
\keyword{internal}
