% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_msigdb_xml}
\alias{import_msigdb_xml}
\title{Import MSigDB XML file}
\usage{
import_msigdb_xml(path)
}
\arguments{
\item{path}{path to file}
}
\value{
gene set contents and information
}
\description{
Import MSigDB XML file
}
\examples{
file <- system.file("extdata", "ex_msig.xml", package = "glacier")
data_pre <- import_msigdb_xml(file)
}
