% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
plot_overlap(matches, value, input, stats, val_trans = "identity")
}
\arguments{
\item{matches}{value from \code{\link{compute}}}

\item{value}{\code{"Gene Value"} for values from \code{input} or one of
\code{"#gene sets"}, \code{"# genes"}, \code{"# matches"},
\code{"P-value"}, \code{"Adjusted P-value"}, \code{"Odds Ratio"},
\code{"Fold Enrichment"} or \code{"Adjusted Fold Enrichment"} for values
from \code{stats}}

\item{input}{output of \code{\link{process_input}}}

\item{stats}{value from \code{\link{compute}}}

\item{val_trans}{optional: value transformation, see \code{trans} argument in
\code{\link[ggplot2]{scale_continuous}}; default \code{"identity"}}
}
\value{
ggplot2: heatmap of overlap
}
\description{
Heatmap will plot genes on the X-axis against annotations on the Y-axis.
Matches will be shown with colored squares, and non-matches will be
transparent. If the \code{value} is "Gene Value", "Odds Ratio", "Fold
Enrichment" or "Adjusted Fold Enrichment", if the value is NaN or an
infinity, these will also be transparent.
}
\examples{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", FALSE, c(2, 10), 11)
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
results <- compute(input, anno, data, 10000)
over <- plot_overlap(results$matches, "Gene Value", input, results$stats)
}
