% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot calculated scores}
\usage{
plot_scores(scores, x, y, color, mode)
}
\arguments{
\item{scores}{value from \code{\link{score_seurat}} or
\code{\link{score_expr}}}

\item{x}{grouping variable e.g. "bin"}

\item{y}{plotting variable e.g. "pca"}

\item{color}{coloring variable e.g. "grp"}

\item{mode}{'whiskers', 'box' or 'violin'}
}
\value{
ggplot2: scores
}
\description{
Plot calculated scores
}
\examples{
seu_path <- system.file("extdata", "ex_seurat.rds", package = "glacier")
seurat <- readRDS(seu_path)

results <- score_seurat(seurat, "grp", c("APOE", "CTSZ"))
scores <- plot_scores(results$scores, "seurat_clusters", "pca", "grp", "box")
}
