\name{plotNDSI}
\alias{plotNDSI}
\title{Plot an NDSI map}
\description{
Displays an NDSI raster.
}
\usage{
plotNDSI(ndsi,
         palette = mako(200),
         title = "NDSI")
}
\arguments{
  \item{ndsi}{Raster generated by glacierNDSI().}
  \item{palette}{Color palette.}
  \item{title}{Plot title.}
}
\value{
Displays an NDSI map.
}
\examples{
\dontrun{
plotNDSI(ndsi)
}
}
\keyword{plot}
