\name{plotNDWI}
\alias{plotNDWI}
\title{Plot an NDWI map}
\description{
Displays an NDWI raster.
}
\usage{
plotNDWI(ndwi,
         palette = viridis(200),
         title = "NDWI")
}
\arguments{
  \item{ndwi}{Raster generated by glacierNDWI().}
  \item{palette}{Color palette.}
  \item{title}{Plot title.}
}
\value{
Displays an NDWI map.
}
\examples{
plotNDWI(ndwi)
}
\keyword{plot}
