% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.annotations}
\alias{process.annotations}
\title{Process annotations}
\usage{
process.annotations(anno.pre, info, options)
}
\arguments{
\item{anno.pre}{output of \code{\link{import.annotations}}}

\item{info}{tibble of gene set name 'name' and information 'info'}

\item{options}{'name' for gene set names, 'syms' for gene set symbols,
'info' for gene set information, 'auto' for automatically generated
annotations, 'file' for manual annotations}
}
\value{
tibble mapping gene sets to annotations and vector of annotations
}
\description{
Process annotations
}
\examples{
\dontrun{
process.annotations(anno.pre, info.pre, c('file', 'auto'))
}
}
