% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.input.post}
\alias{process.input.post}
\title{Post-process gene string input}
\usage{
process.input.post(input.pre)
}
\arguments{
\item{input.pre}{output of \code{\link{process.input.pre}}}
}
\value{
gene names, values and bitmap index of missing values
}
\description{
Post-process gene string input
}
\examples{
\dontrun{
input.pre <- process.input.pre('GENE1 GENE2 GENE3')
process.input.post(input.pre)
}
}
