% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.input.pre}
\alias{process.input.pre}
\title{Pre-process gene string input}
\usage{
process.input.pre(input.raw)
}
\arguments{
\item{input.raw}{string input with genes and values}
}
\value{
pre-processed input
}
\description{
Removes duplicate genes. If duplicate values are found, the first value will
be used.
}
\examples{
\dontrun{
process.input.pre('GENE1 GENE2 GENE3')
process.input.pre('GENE1 0.1 GENE2 0.2 GENE3 0.3')
}
}
