% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.input}
\alias{process.input}
\title{Process gene string input}
\usage{
process.input(input.raw)
}
\arguments{
\item{input.raw}{string input with genes and values}
}
\value{
gene names, values and bitmap index of missing values
}
\description{
Process gene string input
}
\examples{
process.input('GENE1 GENE2 GENE3')
process.input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
}
