% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations}
\usage{
process_annotations(anno, info, options)
}
\arguments{
\item{anno}{output of \code{\link{import_annotations}}}

\item{info}{tibble: "name" gene set name "info" gene set descriptions}

\item{options}{\code{"name"} for gene set names, \code{"syms"} for gene set
symbols, \code{"info"} for gene set descriptions, \code{"auto"} for
automatically generated annotations and/or \code{"file"} for manual
annotations}
}
\value{
\code{gs_annos} tibble: gene sets and annotations

\code{annos} vector: annotations
}
\description{
Choose annotations to include in annotation list, from gene set names, gene
set symbols, gene set descriptions, annotations automatically extracted from
gene set descriptions or manual annotations.
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", FALSE, c(2, 10), 11)
info <- anno[c("name", "info")]

anno_proc <- process_annotations(anno, info, "file")
}
}
\keyword{internal}
