% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_database}
\alias{process_database}
\title{Process database}
\usage{
process_database(data_pre, categories = NULL, organisms = NULL)
}
\arguments{
\item{data_pre}{output of \code{\link{import_database}} or
\code{\link{import_msigdb_xml}}}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}
}
\value{
list of gene sets to genes, gene set information and vector of genes
}
\description{
Process database
}
\examples{
\dontrun{
data_raw <- read_common('path/to/data.csv', ',', T)
data_pre <- import_database(data_raw, c(2, 9), 10)
process_database(data_pre, 'Not assigned', 'Not assigned')
}
}
