% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input}
\alias{process_input}
\title{Process gene string input}
\usage{
process_input(input_raw)
}
\arguments{
\item{input_raw}{string input with genes and values}
}
\value{
gene names, values and bitmap index of missing values
}
\description{
Process gene string input
}
\examples{
process_input('GENE1 GENE2 GENE3')
process_input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
}
