% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input_post}
\alias{process_input_post}
\title{Post-process gene string input}
\usage{
process_input_post(input_pre)
}
\arguments{
\item{input_pre}{output of \code{\link{process_input_pre}}}
}
\value{
gene names, values and bitmap index of missing values
}
\description{
Post-process gene string input
}
\examples{
\dontrun{
input_pre <- process_input_pre('GENE1 GENE2 GENE3')
process_input_post(input_pre)
}
}
