% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input_pre}
\alias{process_input_pre}
\title{Pre-process gene string input}
\usage{
process_input_pre(input_raw)
}
\arguments{
\item{input_raw}{string input with genes and values}
}
\value{
pre-processed input
}
\description{
Removes duplicate genes. If duplicate values are found, the first value will
be used.
}
\examples{
\dontrun{
input_pre <- process_input_pre('CYP1A1 CYP1B1 NQO1 SODD')
input_pre <- process_input_pre('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
}
}
