% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input_seurat}
\alias{process_input_seurat}
\title{Extract differentially expressed genes from Seurat object}
\usage{
process_input_seurat(seurat, clst_1, clst_2 = NULL, max_p = 0.05)
}
\arguments{
\item{seurat}{Seurat object}

\item{clst_1}{first cluster}

\item{clst_2}{optional: second cluster; default all others}

\item{max_p}{P-value cutoff, only genes with P-values less than this will be
returned}
}
\value{
tibble: "gene" gene names "value" gene values
}
\description{
Finds differentially expressed genes, records adjusted P-value and filters
for values less than \code{max_p}.
}
\examples{
seu_path <- system.file("extdata", "ex_seurat.rds", package = "glacier")
seurat <- readRDS(seu_path)

input <- process_input_seurat(seurat, 0)
}
