% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input_text}
\alias{process_input_text}
\title{Process text input}
\usage{
process_input_text(text)
}
\arguments{
\item{text}{character: input with genes and optionally values}
}
\value{
tibble: "gene" gene names "value" gene values
}
\description{
Removes duplicate genes. If multiple values for the same gene are found, only
the first value will be kept.
}
\examples{
input <- process_input_text("FCN1 FTL CLU")
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
}
