% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read.common}
\alias{read.common}
\title{Read delimited files into tibble}
\usage{
read.common(path, delim, header)
}
\arguments{
\item{path}{path to file}

\item{delim}{file delimiter}

\item{header}{whether file contains header}
}
\value{
raw file contents as tibble
}
\description{
Read delimited files into tibble
}
\examples{
\dontrun{
read.common('path/to/common.csv', ',', T)
}
}
