% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_common}
\alias{read_common}
\title{Read delimited files into tibble}
\usage{
read_common(path, delim, header)
}
\arguments{
\item{path}{path to file}

\item{delim}{file delimiter}

\item{header}{whether file contains header}
}
\value{
raw file contents as tibble
}
\description{
Read delimited files into tibble
}
\examples{
file <- system.file("extdata", "ex_anno.csv", package = "glacier")
common_raw <- read_common(file, ",", FALSE)
}
