% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_seurat}
\alias{score_seurat}
\title{Calculate Seurat scores}
\usage{
score_seurat(seurat, group, genes)
}
\arguments{
\item{seurat}{Seurat object}

\item{group}{grouping variable from Seurat data}

\item{genes}{genes of interest}
}
\value{
Seurat pca, pls and plsda scores with AUC
}
\description{
Calculate Seurat scores
}
\examples{
seu_path <- system.file("extdata", "ex_seurat.rds", package = "glacier")
seurat <- readRDS(seu_path)

results <- score_seurat(seurat, "grp", c("APOE", "CTSZ"))
}
