% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{uses}
\alias{uses}
\title{Checks that package can be used, else returns error}
\usage{
uses(packages, error, message)
}
\arguments{
\item{packages}{character vector: packages required}

\item{error}{error function}

\item{message}{error message}
}
\description{
Checks that package can be used, else returns error
}
\examples{
\dontrun{
uses("stringr", stop, "Prerequisite not available")
}
}
\keyword{internal}
