% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Calculate overlap statistics for select genes}
\usage{
workflow(
  input.raw,
  anno.pre,
  anno.opts = "file",
  data.pre,
  info.from = "database",
  categories = NULL,
  organisms = NULL,
  universe
)
}
\arguments{
\item{input.raw}{string input with genes and values}

\item{anno.pre}{output of \code{\link{import.annotations}}}

\item{anno.opts}{annotation types to use}

\item{data.pre}{output of \code{\link{import.database}}}

\item{info.from}{"annotation" or "database"}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Calculate overlap statistics for select genes
}
\examples{
\dontrun{
workflow(input.raw, anno.pre, data.pre, 10000)
}
}
