% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{annotation.map}
\alias{annotation.map}
\title{Map annotation to gene sets and genes}
\usage{
annotation.map(annotation, gs_annos, gs_genes)
}
\arguments{
\item{annotation}{annotation to map}

\item{gs_annos}{value from \code{\link{process.annotations}}}

\item{gs_genes}{value from \code{\link{process.database}}}
}
\value{
vector of associated gene sets and vector of associated genes
}
\description{
Map annotation to gene sets and genes
}
\examples{
\dontrun{
annotation.map('annotation', gs_annos, gs_genes)
}
}
