% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate.post}
\alias{calculate.post}
\title{Finish statistical calculations}
\usage{
calculate.post(stats.pre, input.count, universe)
}
\arguments{
\item{stats.pre}{value from \code{\link{calculate.pre}}}

\item{input.count}{number of genes in input}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Finish statistical calculations
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
anno.pre <- import.annotations(anno.raw, c(2, 9), 10)
info.pre <- anno.pre[c('name', 'info')]
anno <- process.annotations(anno.pre, info.pre, c('file', 'auto'))

data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)
data <- process.database(data.pre, 'Not assigned', 'Not assigned')

input <- process.input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
calc.pre <- calculate.pre(input, anno$annos, anno$gs_annos, data$gs_genes)
calculate.post(calc.pre$stats.pre, nrow(input), 10000)
}
}
