% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, annos, gs_annos, gs_genes, universe)
}
\arguments{
\item{input}{value from \code{\link{process_input}}}

\item{annos}{value from \code{\link{process_annotations}}}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}

\item{universe}{number of genes in universe}
}
\value{
tibble of overlap statistics and list of annotation matches
}
\description{
Calculate overlap statistics
}
\examples{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
calc <- calculate(input$input, anno$annos, anno$gs_annos, data$gs_genes, 100)
}
