% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, annos, gs_annos, gs_genes, universe)
}
\arguments{
\item{input}{value from \code{\link{process.input}}}

\item{annos}{value from \code{\link{process.annotations}}}

\item{gs_annos}{value from \code{\link{process.annotations}}}

\item{gs_genes}{value from \code{\link{process.database}}}

\item{universe}{number of genes in universe}
}
\value{
tibble of overlap statistics and list of annotation matches
}
\description{
Calculate overlap statistics
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
anno.pre <- import.annotations(anno.raw, c(2, 9), 10)
info.pre <- anno.pre[c('name', 'info')]
anno <- process.annotations(anno.pre, info.pre, c('file', 'auto'))

data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)
data <- process.database(data.pre, 'Not assigned', 'Not assigned')

input <- process.input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
calculate(input$input, anno$annos, anno$gs_annos, data$gs_genes, 10000)
}
}
