% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, annos, gs_annos, gs_genes, universe = NULL)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{annos}{value from \code{\link{process_annotations}}}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}

\item{universe}{optional: number of genes in universe; default calculate from
\code{gs_genes}}
}
\value{
\code{stats} tibble: overlap statistics

\code{matches} list: names: annotations vector: matched genes
}
\description{
Calculate overlap statistics
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", TRUE, c(2, 4), 5)
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)
info <- anno[c("name", "info")]

anno_proc <- process_annotations(anno, info, "file")
data_proc <- process_database(data, "Not assigned", "Not assigned")
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
calc <- calculate(input, anno_proc$annos, anno_proc$gs_annos,
                  data_proc$gs_genes)
}
}
\keyword{internal}
