% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, annos, gs_annos, gs_genes, universe)
}
\arguments{
\item{input}{output of \code{\link{process_input}}}

\item{annos}{value from \code{\link{process_annotations}}}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}

\item{universe}{optional: number of genes in universe; default calculate from
\code{gs_genes}}
}
\value{
\code{stats} tibble: overlap statistics

\code{matches} list: names: annotations vector: matched genes
}
\description{
Calculate overlap statistics
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", FALSE, c(2, 10), 11)
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

info <- anno[c("name", "info")]
anno_proc <- process_annotations(anno, info, "file")
data_proc <- process_database(data, "Not assigned", "Not assigned")

input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
calc <- calculate(input, anno_proc$annos, anno_proc$gs_annos,
                  data_proc$gs_genes, 10000)
}
}
\keyword{internal}
