% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_post}
\alias{calculate_post}
\title{Finish statistical calculations}
\usage{
calculate_post(stats_pre, input_count, universe)
}
\arguments{
\item{stats_pre}{value from \code{\link{calculate_pre}}}

\item{input_count}{number of genes in input}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Finish statistical calculations
}
\examples{
\dontrun{
anno_raw <- read_common('path/to/anno.csv', ',', T)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)
info_pre <- anno_pre[c('name', 'info')]
anno <- process_annotations(anno_pre, info_pre, c('file', 'auto'))

data_raw <- read_common('path/to/data.csv', ',', T)
data_pre <- import_database(data_raw, c(2, 9), 10)
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
calc_pre <- calculate_pre(input, anno$annos, anno$gs_annos, data$gs_genes)
calculate_post(calc_pre$stats_pre, nrow(input), 10000)
}
}
