% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_pre}
\alias{calculate_pre}
\title{Begin calculating overlap statistics}
\usage{
calculate_pre(input, annos, gs_annos, gs_genes)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{annos}{value from \code{\link{process_annotations}}}

\item{gs_annos}{value from \code{\link{process_annotations}}}

\item{gs_genes}{value from \code{\link{process_database}}}
}
\value{
\code{stats_pre} tibble: overlap statistics (incomplete)

\code{matches} list: names: annotations vector: matched genes
}
\description{
Begin calculating overlap statistics
}
\examples{
\dontrun{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", TRUE, c(2, 4), 5)
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)
info <- anno[c("name", "info")]

anno_proc <- process_annotations(anno, info, "file")
data_proc <- process_database(data, "Not assigned", "Not assigned")
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
calc_pre <- calculate_pre(input, anno_proc$annos, anno_proc$gs_annos,
                          data_proc$gs_genes)
}
}
\keyword{internal}
