% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display.overlap}
\alias{display.overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
display.overlap(input, stats, matches, value, value.trans = "identity")
}
\arguments{
\item{input}{value from \code{\link{process.input}}}

\item{stats}{value from \code{\link{calculate}}}

\item{matches}{value from \code{\link{calculate}}}

\item{value}{option for tile color}

\item{value.trans}{optional value transformation}
}
\value{
ggplot heatmap
}
\description{
Plot gene-annotation matches as heatmap
}
\examples{
\dontrun{
display.overlap(input, stats, matches, 'Gene Value')
}
}
