% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display.stats}
\alias{display.stats}
\title{Plot overlap statistics on a bar graph}
\usage{
display.stats(
  stats,
  value,
  color,
  value.trans = "identity",
  color.trans = "identity",
  sort = F
)
}
\arguments{
\item{stats}{value from \code{\link{calculate}}}

\item{value}{option for bar length}

\item{color}{option for bar color}

\item{value.trans}{optional value transformation}

\item{color.trans}{optional color transformation}

\item{sort}{whether to sort annotations alphabetically}
}
\value{
ggplot bar graph
}
\description{
Plot overlap statistics on a bar graph
}
\examples{
\dontrun{
display.stats(stats, value, color)
}
}
