% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_overlap}
\alias{display_overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
display_overlap(input, stats, matches, value, value_trans = "identity")
}
\arguments{
\item{input}{value from \code{\link{process_input}}}

\item{stats}{value from \code{\link{calculate}}}

\item{matches}{value from \code{\link{calculate}}}

\item{value}{option for tile color as 'Gene Value' or one of '# gene sets',
'#genes', '# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio',
'Fold Enrichment' or 'Adjusted Fold Enrichment'}

\item{value_trans}{optional value transformation}
}
\value{
ggplot heatmap
}
\description{
Plot gene-annotation matches as heatmap
}
\examples{
\dontrun{
anno_raw <- read.common('path/to/anno.csv', ',', T)
anno_pre <- import.annotations(anno_raw, c(2, 9), 10)
info <- anno_pre[c('name', 'info')]
anno <- process_annotations(anno_pre, info, c('file', 'auto'))

data_raw <- read.common('path/to/data.csv', ',', T)
data_pre <- import.database(data_raw, c(2, 9), 10)
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
stats <- calculate(input, anno$annos, anno$gs_annos, data$gs_genes, 10000)
display_overlap(input$gene, stats$stats, stats$matches, 'Gene Value')
}
}
