% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_stats}
\alias{display_stats}
\title{Plot overlap statistics on a bar graph}
\usage{
display_stats(
  stats,
  value,
  color,
  value_trans = "identity",
  color_trans = "identity",
  sort = F
)
}
\arguments{
\item{stats}{value from \code{\link{calculate}}}

\item{value}{option for bar length as one of '# gene sets', '#genes',
'# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio', 'Fold Enrichment'
or 'Adjusted Fold Enrichment'}

\item{color}{option for bar color as one of '# gene sets', '#genes',
'# matches', 'P-value', 'Adjusted P-value', 'Odds Ratio', 'Fold Enrichment'
or 'Adjusted Fold Enrichment'}

\item{value_trans}{optional value transformation}

\item{color_trans}{optional color transformation}

\item{sort}{whether to sort annotations alphabetically}
}
\value{
ggplot bar graph
}
\description{
Plot overlap statistics on a bar graph
}
\examples{
\dontrun{
anno_raw <- read.common('path/to/anno.csv', ',', T)
anno_pre <- import.annotations(anno_raw, c(2, 9), 10)
info <- anno_pre[c('name', 'info')]
anno <- process_annotations(anno_pre, info, c('file', 'auto'))

data_raw <- read.common('path/to/data.csv', ',', T)
data_pre <- import.database(data_raw, c(2, 9), 10)
data <- process_database(data_pre, 'Not assigned', 'Not assigned')

input <- process_input('GENE1 0.1 GENE2 0.2 GENE3 0.3')
stats <- calculate(input, anno$annos, anno$gs_annos, data$gs_genes, 10000)
display_stats(stats$stats, 'Fold Enrichment', 'Adjusted P-value')
}
}
