% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.annotations}
\alias{import.annotations}
\title{Import annotation file}
\usage{
import.annotations(common.raw, cols = NULL, info = NULL)
}
\arguments{
\item{common.raw}{raw file tibble or output of \code{\link{import.common}}}

\item{cols}{data column boundaries}

\item{info}{info column}
}
\value{
gene set annotations and information
}
\description{
Import annotation file
}
\examples{
\dontrun{
import.annotations(raw, c(2, 9), 10)
}
}
