% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.common}
\alias{import.common}
\title{Pre-process common file type tibble}
\usage{
import.common(common.raw, cols, info)
}
\arguments{
\item{common.raw}{raw file tibble}

\item{cols}{data column boundaries}

\item{info}{info column}
}
\value{
pre-processed common file type tibble
}
\description{
Pre-process common file type tibble
}
\examples{
\dontrun{
import.common(raw, c(2, 9), 10)
}
}
