% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_annotations}
\alias{import_annotations}
\title{Import annotation file}
\usage{
import_annotations(anno_raw, cols = NULL, info = NULL)
}
\arguments{
\item{anno_raw}{raw file tibble or output of \code{\link{import_common}}}

\item{cols}{data column boundaries}

\item{info}{info column}
}
\value{
gene set annotations and information
}
\description{
Import annotation file
}
\examples{
\dontrun{
anno_raw <- read_common('path/to/anno.csv', ',', T)
import_annotations(anno_raw, c(2, 9), 10)
}
}
