% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_common}
\alias{import_common}
\title{Pre-process common file type tibble}
\usage{
import_common(common_raw, cols, info)
}
\arguments{
\item{common_raw}{raw file tibble}

\item{cols}{data column boundaries}

\item{info}{info column}
}
\value{
pre-processed common file type tibble
}
\description{
Pre-process common file type tibble
}
\examples{
\dontrun{
common_raw <- read_common('path/to/common.csv', ',', T)
import_common(common_raw, c(2, 9), 10)
}
}
