% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_database}
\alias{import_database}
\title{Import database file}
\usage{
import_database(data_raw, data = NULL, info = NULL)
}
\arguments{
\item{data_raw}{raw file tibble or output of \code{\link{import_common}}}

\item{data}{data column boundaries}

\item{info}{info column}
}
\value{
gene set contents and information
}
\description{
Import database file
}
\examples{
\dontrun{
data_raw <- read.common('path/to/data.csv', ',', T)
import_database(data_raw, c(2, 9), 10)
}
}
