% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_database}
\alias{import_database}
\title{Import database file}
\usage{
import_database(data_raw, data = NULL, info = NULL)
}
\arguments{
\item{data_raw}{raw file tibble or output of \code{\link{import_common}}}

\item{data}{data column boundaries}

\item{info}{info column}
}
\value{
gene set contents and information
}
\description{
Import database file
}
\examples{
file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)
}
