% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot gene-annotation matches as heatmap}
\usage{
plot_overlap(matches, value, input, stats, xpos = "bottom")
}
\arguments{
\item{matches}{value from \code{\link{compute}}}

\item{value}{\code{"Gene Value"} for values from \code{input} or one of
\code{"#gene sets"}, \code{"# genes"}, \code{"# matches"},
\code{"P-value"}, \code{"Adjusted P-value"}, \code{"Odds Ratio"},
\code{"Fold Enrichment"} or \code{"Adjusted Fold Enrichment"} for values
from \code{stats}}

\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{stats}{value from \code{\link{compute}}}

\item{xpos}{x axis position, "top" or "bottom"}
}
\value{
ggplot2: heatmap of overlap
}
\description{
Heatmap will plot genes on the X-axis against annotations on the Y-axis.
Matches will be shown with colored squares, and non-matches will be
transparent. If the \code{value} is "Gene Value", "Odds Ratio", "Fold
Enrichment" or "Adjusted Fold Enrichment", if the value is NaN or an
infinity, these will also be transparent.
}
\examples{
anno_path <- system.file("extdata", "ex_anno.csv", package = "glacier")
data_path <- system.file("extdata", "ex_data.csv", package = "glacier")
anno <- import_annotations(anno_path, ",", TRUE, c(2, 4), 5)
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
results <- compute(input, anno, data)
over <- plot_overlap(results$matches, "Gene Value", input, results$stats)
}
