% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.database}
\alias{process.database}
\title{Process database}
\usage{
process.database(data.pre, categories = NULL, organisms = NULL)
}
\arguments{
\item{data.pre}{output of \code{\link{import.database}} or
\code{\link{import.msigdb_xml}}}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}
}
\value{
list of gene sets to genes, gene set information and vector of genes
}
\description{
Process database
}
\examples{
\dontrun{
data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)
process.database(data.pre, 'Not assigned', 'Not assigned')
}
}
