% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations}
\usage{
process_annotations(anno_pre, info, options)
}
\arguments{
\item{anno_pre}{output of \code{\link{import_annotations}}}

\item{info}{tibble of gene set name 'name' and information 'info'}

\item{options}{'name' for gene set names, 'syms' for gene set symbols,
'info' for gene set information, 'auto' for automatically generated
annotations, 'file' for manual annotations}
}
\value{
tibble mapping gene sets to annotations and vector of annotations
}
\description{
Process annotations
}
\examples{
file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

info <- anno_pre[c("name", "info")]
anno <- process_annotations(anno_pre, info, "file")
}
