% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_database}
\alias{process_database}
\title{Process database}
\usage{
process_database(data_pre, categories = NULL, organisms = NULL)
}
\arguments{
\item{data_pre}{output of \code{\link{import_database}} or
\code{\link{import_msigdb_xml}}}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}
}
\value{
list of gene sets to genes, gene set information and vector of genes
}
\description{
Process database
}
\examples{
file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

data <- process_database(data_pre, 'Not assigned', 'Not assigned')
}
