% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input}
\alias{process_input}
\title{Process gene string input}
\usage{
process_input(input_raw)
}
\arguments{
\item{input_raw}{string input with genes and values}
}
\value{
gene names, values and bitmap index of missing values
}
\description{
Process gene string input
}
\examples{
input <- process_input('CYP1A1 CYP1B1 NQO1 SODD')
input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
}
