% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_input}
\alias{process_input}
\title{Process text input}
\usage{
process_input(text)
}
\arguments{
\item{text}{character: input with genes and optionally values}
}
\value{
tibble: "gene" gene names "value" gene values
}
\description{
Removes duplicate genes. If multiple values for the same gene are found, only
the first value will be kept.
}
\examples{
input <- process_input('CYP1A1 CYP1B1 NQO1 SODD')
input <- process_input('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0')
}
