% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Calculate overlap statistics for select genes}
\usage{
workflow(
  input.raw,
  anno.pre,
  anno.opts = "file",
  data.pre,
  info.from = "database",
  categories = NULL,
  organisms = NULL,
  universe
)
}
\arguments{
\item{input.raw}{string input with genes and values}

\item{anno.pre}{output of \code{\link{import.annotations}}}

\item{anno.opts}{'name' for gene set names, 'syms' for gene set symbols,
'info' for gene set information, 'auto' for automatically generated
annotations, 'file' for manual annotations}

\item{data.pre}{output of \code{\link{import.database}}}

\item{info.from}{"annotation" or "database"}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Calculate overlap statistics for select genes
}
\examples{
\dontrun{
anno.raw <- read.common('path/to/anno.csv', ',', T)
anno.pre <- import.annotations(anno.raw, c(2, 9), 10)

data.raw <- read.common('path/to/data.csv', ',', T)
data.pre <- import.database(data.raw, c(2, 9), 10)

workflow('GENE1 0.1 GENE2 0.2 GENE3 0.3', anno.pre, data.pre = data.pre,
         universe = 10000)
}
}
