% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Calculate overlap statistics for select genes}
\usage{
workflow(
  input_raw,
  anno_pre,
  anno_opts = "file",
  data_pre,
  info_from = "database",
  categories = NULL,
  organisms = NULL,
  universe
)
}
\arguments{
\item{input_raw}{string input with genes and values}

\item{anno_pre}{output of \code{\link{import_annotations}}}

\item{anno_opts}{'name' for gene set names, 'syms' for gene set symbols,
'info' for gene set information, 'auto' for automatically generated
annotations, 'file' for manual annotations}

\item{data_pre}{output of \code{\link{import_database}}}

\item{info_from}{"annotation" or "database"}

\item{categories}{optionally filter categories}

\item{organisms}{optionally filter organisms}

\item{universe}{number of genes in universe}
}
\value{
overlap statistics
}
\description{
Calculate overlap statistics for select genes
}
\examples{
anno_file <- system.file("extdata", "ex_anno.csv", package = "glacier")
anno_raw <- read_common(anno_file, ",", FALSE)
anno_pre <- import_annotations(anno_raw, c(2, 9), 10)

data_file <- system.file("extdata", "ex_data.csv", package = "glacier")
data_raw <- read_common(data_file, ",", FALSE)
data_pre <- import_database(data_raw, c(2, 4), 0)

calc <- workflow('CYP1A1 0.2 CYP1B1 NQO1 0.3 SODD 9.0', anno_pre,
                 data_pre = data_pre, universe = 100)
}
